backend/requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6
PyPDF2>=3.0.0
openpyxl>=3.1.0
python-dotenv>=1.0.0
supabase>=2.0.0
anthropic>=0.7.0
pydantic>=2.5.0
pydantic-settings>=2.0.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
